Forecasting Task (Regression)

This task specializes Task and TaskSupervised for forecasting regression problems. The target column(s) are assumed to be numeric. The task_type is set to "regr" `.

Format

R6::R6Class object inheriting from Task/TaskSupervised/TaskForecast.

Construction

t = TaskRegrForecast$new(id, backend, target, date_col)

id :: character(1)
Identifier for the task.

backend :: DataBackend
Either a DataBackendLong, a object of class ts or a data.frame with specified date column

or any object which is convertible to a DataBackend with as_data_backend(). E.g., a object of class dts will be converted to a DataBackendLong.

target :: character(n)
Name of the target column(s).

Fields

All methods from TaskSupervised and TaskForecast.

Methods

See TaskSupervised and TaskForecast.

See also

Super classes

mlr3::Task -> mlr3::TaskSupervised -> mlr3temporal::TaskForecast -> TaskRegrForecast

Methods

Public methods

TaskRegrForecast$new()

TaskRegrForecast$clone()

Inherited methods

mlr3::Task$add_strata()

mlr3::Task$cbind()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$print()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_levels()

mlr3::Task$set_row_roles()

mlr3temporal::TaskForecast$data()

mlr3temporal::TaskForecast$date()

mlr3temporal::TaskForecast$truth()

Method new()

Usage

TaskRegrForecast$new(id, backend, target = "target", date_col = NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskRegrForecast$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.