Vector Autoregression Forecast Learner

Vector autoregressive model Calls vars::VAR from package vars.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("forecast.VAR")
lrn("forecast.VAR")

Meta Information

Task type: “forecast”

Predict Types: “response”, “se”

Feature Types: “numeric”

Required Packages: mlr3, vars

Parameters

Id	Type	Default	Range

p	integer	1	\([0, \infty)\)

lag.max	integer	NULL	\([1, \infty)\)

season	integer	NULL	\([1, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3temporal::LearnerForecast -> LearnerVAR

Methods

Public methods

LearnerRegrForecastVAR$new()

LearnerRegrForecastVAR$forecast()

LearnerRegrForecastVAR$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3temporal::LearnerForecast$fitted_values()

mlr3temporal::LearnerForecast$predict()

mlr3temporal::LearnerForecast$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrForecastVAR$new()

Method forecast()

Returns forecasts after the last training instance.

Usage

LearnerRegrForecastVAR$forecast(h = 10, task, new_data = NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrForecastVAR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("forecast.VAR")
print(learner)
#> <LearnerVAR:forecast.VAR>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, vars
#> * Predict Types:  [response], se
#> * Feature Types: numeric
#> * Properties: exogenous, missings, multivariate

# available parameters:
learner$param_set$ids()
#> [1] "p"       "lag.max" "season"