Forecasting Task

Introduces a new Forecasting Task that inherits from TaskSupervised which can be used to forecast time series. This is an abstract class, which should not be instantiated. Use the subclass TaskRegrForecast instead. The task_type is set to "forecast".

Note, that in case the input is a data.table' or data.frame, mlr3temporalexpects a "wide" data.frame as input. Thetsbox::to_wide()` function can help casting time-series to this format.

Format

R6::R6Class object inheriting from Task/TaskSupervised.

Construction

t = TaskForecast$new(id, backend, target, date_col)

id :: character(1)
Identifier for the task.

backend :: DataBackend
Either a DataBackendLong, an object of class ts or a data.frame with specified date column or any object which is convertible to a DataBackendLong with as_data_backend(). E.g., a object of class dts (from package tsbox) will be converted to a DataBackendLong.

target :: character(n)
Name of the target column(s).

date_col :: character(1)
Name of the date column, only required if backend is a data.frame.

Fields

All methods inherited from TaskSupervised, and additionally:

date_col :: character(1)
Name of the date column.

Methods

See TaskSupervised, additionally:

date(row_ids = NULL) :: data.table
(integer() | character()) -> named list()
Returns the date column.

See also